Disable `--rerun-fails` in tests
When `--rerun-fails` is specified and a test is re-run, the produced coverage data only contains the coverage from the failed but re-executed sub-tests.

This change sees how we fare with this disabled as we've recently addressed some of the test flakiness. We'll be relying on the much less granular `./scripts/retry` for retries, which doesn't have the coverage problem. The downside is it will re-run all the tests.