[sdk/python] Install local SDK in automation API test (#15222)
In preparation for supporting Python 3.12...

The test was installing a `requirements.txt` file that depends on the
released version of the `pulumi` package, but that package depends on
the old `grpcio` package that does not work Python 3.12. Instead of
depending on the public `pulumi` package, install the locally built
Python SDK.

Also, fix the test to be able to run on Windows, while making changes
here.