reduce the number of partitions for macos runners
I was watching the merge queue quite closely today, and it's
continually stuck waiting to run jobs on MacOS runners.  However when
the jobs are running on them they tend to be faster than e.g. Windows
runners, often running the tests assigned to them in 2 min, vs. 8 min
on windows.

In addition we're spending quite a bit of time to setup the actual
running of tests, and by spreading the tests out over multiple runners
that time multiplies.  Given the macos runners are the most expensive
ones that probably also adds up.

Given all that, we can reduce the number of MacOS runners we're using,
while not trading that off against longer overall CI times, as those
are limited by slower jobs.

Unfortunately this results in some duplication in the ci job, but I
think the tradeoff here might be worth it.