Temporarily skip more python smoke tests
These tests depends on the published version of `pulumi` which can't be installed on Python 3.12 due to it requiring an older version of `grpcio` which can't be installed on Python 3.12.

For now, temporarily skip these tests. After we've released a new version of `pulumi` that can be installed on Python 3.12, we can re-enable the test.