upgrade grpc-js to latest release (#15231)
We've previously pinned this to 1.9.6, because there was a bug that
caused an automation API hang in 1.9.7. There's been a few patch
releases since, so let's try to upgrade again, as this could potentially
fix some flaky
tests (https://github.com/pulumi/pulumi/issues/14841).

The PR doing the downgrade originally also introduced a test, so as long
as that test is passing upgrading should be safe.

Fixes https://github.com/pulumi/pulumi/issues/14841