[sdk/go] Use `filepath.Join` rather than `path.Join` (#15273)
We're seeing Go Automation API test failures on Windows. Looks like it's
likely due to the use of `path.Join` (introduced in
https://github.com/pulumi/pulumi/pull/15049 which merged about an hour
ago; it's a separate question on how it was able to merge with failures
on Windows). This PR changes those uses of `path` to `filepath`.

Note: No changelog because this hasn't been released yet.

This is currently blocking merges.

Fixes #15272