Test main in conformance tests
This adds a main option to the test, that is put into the Pulumi.yaml
and sent to codegen.  This allows us to write a test that sets up "main" in a
Pulumi.yaml file and test that "program directory" and "entry point" are
respected. Doing this correctly also required some bug fixes in the conformance
interface itself.