ci: remove unnecessary TODO
This TODO has been there since this 6ef7ef64d620dafda126e1623c92ea2ff4823031
was merged to the main branch, but the flag was always true, and the
comment also tells us to set it to true.  Setting it to true is still
the right thing, so let's just get rid of the confusing comment.