Don't mutate RegisterResourceRequest (#15297)
Noticed this while trying to prepare a PR for the transforms support.
These tests were checking that we mutated a field on the request object
itself. We no longer do that in the transform branch because everything
has to go through the transform callbacks as well but it caused these
tests to fail.

I've deleted these tests and written a new engine test to validate the
system output is actually as expected.