Unfreeze v3.104.1 (#15335)
I was trying to speed up release of v3.104.1 by combining the change we
wanted to release and the freeze at the same time
(https://github.com/pulumi/pulumi/pull/15333), but that doesn't work. It
left us with a v3.104.2 draft release with the change. I should have
gotten the fix merged, and then followed up with a change to freeze it.

To address, we need to unfreeze and then refreeze.