Auto-fix encrypted keys in the wrong format due to gocloud.dev upgrade regression (#15334)
When attempting another gocloud.dev upgrade, some users ran into a
regression related to the format change of encrypted keys between
gocloud versions, which part of the system was not accounting for. This
PR addresses the issue and includes a fix that automatically fixes
forward state that has an encrypted key in the wrong format, and
includes a regression test and test for the auto-fix behavior.

Fixes #15329

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>