ast: ranges for accessors in single-line flow strings (#231)
These changes add support for reporting ranges for accessors in
single-line flow strings (i.e. single-line strings that are not quoted).

Multi-line strings are quite a bit more complicated to support. The
string we get from the YAML parser has already been processed to remove
indentation, fold newlines, etc., so its bytes do not represent the
bytes in the original file. We will want to add support for these
strings in the future, but doing so is something of an open problem.

Fixes #232.