[sdkgen/python] Require Python >=3.8
Python 3.7 is unsupported and has been end-of-life since 6/27/2023. See https://devguide.python.org/versions/

The core `pulumi` Python SDK package is being updated to indicate that Python 3.8 or greater is required.

This change updates the default minimum required Python version for generated provider SDKs to Python 3.8 or greater as well.