add no_tmp_dir query parameter to file:// URLs
In gocloud.dev 0.34.0 the behaviour for file:// URLs changed, making
pulumi fail when the state is supposed to be written to a mounted
directory.  See also https://github.com/pulumi/pulumi/issues/15352.

Restore th old behaviour by adding a `no_tmp_dir=true` flag to the
URL, unless the user provided it themselves.  We also allow users to
pass `no_tmp_dir=false` in the parameters, which will opt in to the
new behaviour.