add no_tmp_dir query parameter to file:// URLs (#15375)
In gocloud.dev 0.34.0 the behaviour for file:// URLs changed, making
pulumi fail when the state is supposed to be written to a mounted
directory. See also https://github.com/pulumi/pulumi/issues/15352.

Restore the old behaviour by adding a `no_tmp_dir=true` flag to the URL,
unless the user provided it themselves. We also allow users to pass
`no_tmp_dir=false` in the parameters, which will opt in to the new
behaviour.

This actually implements option 2 from the linked issue since it was
easy enough to do, though I don't expect may users to make use of it,
since I don't think a lot of people care about the contents of whatever
directory the state is stored in anyway, so additional temp files there
don't really matter.

Fixes https://github.com/pulumi/pulumi/issues/15352

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->

---------

Co-authored-by: Justin Van Patten <jvp@justinvp.com>