[sdk/python] Upgrade to grpcio 1.60.1 (#15366)
grpcio 1.60.1 includes a fix for the regression that results in a
traceback being written to stderr for unimplemented gRPC methods.
https://github.com/grpc/grpc/releases/tag/v1.60.1

This change upgrades to 1.60.1 and reverts the workarounds we put in
place in https://github.com/pulumi/pulumi/pull/15190.

Note that we have an integration test for Python dynamic providers that
ensures nothing unexpected is written to stderr:
https://github.com/pulumi/pulumi/blob/16a1fb31c2040785dbbfb9de43d772d8e16fffaa/tests/integration/integration_python_acceptance_test.go#L111-L117

Fixes #15365