build resourcedocsgen from registry (#14157)
resourcedocsgen has been remved from the pulumi/docs repo in
https://github.com/pulumi/docs/pull/10009, so now the related CI build
fails. Get it from the pulumi/registry repository instead, where it
lives now.

Note that this requires https://github.com/pulumi/docs/pull/10053 to be
merged first to update the script to no longer try to build the tool.

I'm not sure why the CI job didn't break earlier, the first time I
noticed it being broken was in
https://github.com/pulumi/pulumi/pull/14156.

## Checklist

- [ ] I have run `make tidy` to update any new dependencies
- [ ] I have run `make lint` to verify my code passes the lint check
  - [ ] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [ ] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [ ] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->