Program dependencies aren't necessarily semantic versions
This weakens the contract of GetProgramDependencies that the versions returned
are just "strings". The format of those strings is defined by the language
plugin. This is especially useful for python which uses PEP440 not semantic
versions for it's pip dependencies.

There wasn't anywhere in the system that took a strong dependency on these
being semvers.