Warn about non-project mode (#15411)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

In preparation for deprecating non-project mode at some point. This is
phase 1 of encouraging users off of legacy mode.

Any time a user opens a backend in legacy mode (either old states, or
new states with DIY_BACKEND_LEGACY_LAYOUT set) we'll warn that
non-project mode is due to deprecate this year.

They can disable this warning by setting
PULUMI_DIY_BACKEND_ACKNOWLEDGE_LEGACY_WARNING to true.

The next phase (in a few months) will turn this into an error which
they'll be able to opt out of by setting
PULUMI_DIY_BACKEND_ACKNOWLEDGE_LEGACY_ERROR to true.

After that it will be a hard error and all non-project related code will
be removed.

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [ ] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->